Full attribute parity and result parity between Pardiso and Mumps
It's now possible to use the same solver_opts values in SimPEG for both
the Pardiso and MUMPS solvers.

We now run all the Pardiso test cases for Mumps as well.  Running these
test cases uncovered a Mumps issue where matrix transposes didn't
commute when operating on complex matrices.  If you asked the Mumps
solver:

    Ainv = Mumps(A)
    AinvT_1 = Mumps(A).T
    AinvT_2 = Mumps(A.T)

then the two inverse transposes weren't equal when A is a complex matrix.

For small matrices, you can look at the inverse transposes by solving
this trivial system:

    AinvT_1 * np.identity(A.size) != AinvT_2 * np.identity(A.size)

This issue occurred because pymatsolver treated the complex case as the
conjugate transpose -- but we weren't asking for the conjugate
transpose; we were explicitly asking for just the transpose.

There was no issue with Mumps itself, just a few extra lines of Fortran
code in the Mumps interface.  Now that those lines are removed, the Mumps
solver and the Pardiso solver operate identically with these transposes.